refactor: index.ts 에서 export 하도록 수정
